+++ push_metrics/push.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package metrics
+package push_metrics
 
 import (
 	"context"
